requirements.txt
streamlit
pandas
pandasql
python-dotenv
google-genai
semantic-router[local]
chromadb
sentence_transformers
SQLAlchemy
psycopg2-binary